requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
dataclasses-json==0.5.7
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
faiss-cpu==1.7.4
frozenlist==1.3.3
fuzzywuzzy==0.18.0
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.23.2
ipython==8.12.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter-client==8.2.0
jupyter-core==5.3.1
langchain==0.0.195
langchainplus-sdk==0.0.8
Levenshtein==0.21.1
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
nltk==3.8.1
numexpr==2.8.4
numpy==1.24.3
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
platformdirs==3.5.3
prompt-toolkit==3.0.38
protobuf==4.23.2
psutil==5.9.5
pure-eval==0.2.2
pyarrow==12.0.0
pydantic==1.10.9
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pypdf==3.9.1
pypdf2==3.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-Levenshtein==0.21.1
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==306
PyYAML==6.0
pyzmq==25.1.0
rapidfuzz==3.1.1
regex==2023.6.3
requests==2.31.0
rich==13.4.1
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
smmap==5.0.0
SQLAlchemy==2.0.15
stack-data==0.6.2
streamlit==1.23.1
streamlit-chat==0.0.2.2
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.4.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing-extensions==4.6.3
typing-inspect==0.9.0
tzdata==2023.3
tzlocal==4.3
urllib3==2.0.3
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
yarl==1.9.2
zipp==3.15.0
